+++ radp/digital_twin/rf/bayesian/tests/test_bayesian_engine.py
@@ -9,15 +9,13 @@ is trained and inferred on, using small sample data.
 """
 
 import logging
-import os
-import random
 import unittest
 
 import numpy as np
 import pandas as pd
-import torch
 from sklearn.model_selection import train_test_split
 
+from radp.common.utils import seed_everything
 from radp.digital_twin.rf.bayesian.bayesian_engine import BayesianDigitalTwin
 from radp.digital_twin.utils import constants
 from radp.digital_twin.utils.gis_tools import GISTools
@@ -141,6 +139,90 @@ def get_x_max_and_x_min():
     return x_max, x_min
 
 
+def expected_loss_vs_iter(cell_id: str):
+    """Below test data loss_vs_iter is generated with seed_everything(1)
+    Returns loss_vs_iter for given cell_id"""
+
+    loss_vs_iter = {
+        "Cell1": [
+            1.3360857963562012,
+            1.3250646591186523,
+            1.3147950172424316,
+            1.305358648300171,
+            1.296657919883728,
+            1.288337230682373,
+            1.2800240516662598,
+            1.2716268301010132,
+            1.2632856369018555,
+            1.2552096843719482,
+            1.2475690841674805,
+            1.2404227256774902,
+            1.2336828708648682,
+            1.2271721363067627,
+            1.2207590341567993,
+            1.2144416570663452,
+            1.208317518234253,
+            1.202500820159912,
+            1.197044849395752,
+            1.1918997764587402,
+        ],
+        "Cell2": [
+            1.4958915710449219,
+            1.4850997924804688,
+            1.476243019104004,
+            1.46876859664917,
+            1.461909294128418,
+            1.455451250076294,
+            1.4494956731796265,
+            1.4441474676132202,
+            1.4394338130950928,
+            1.4352850914001465,
+            1.4315435886383057,
+            1.4280192852020264,
+            1.4245692491531372,
+            1.4211323261260986,
+            1.4177082777023315,
+            1.4143176078796387,
+            1.41096830368042,
+            1.4076392650604248,
+            1.404287576675415,
+            1.400879144668579,
+        ],
+        "Cell3": [
+            1.4421265125274658,
+            1.4319207668304443,
+            1.4234235286712646,
+            1.4165308475494385,
+            1.4108211994171143,
+            1.4056122303009033,
+            1.4002456665039062,
+            1.3943250179290771,
+            1.3877692222595215,
+            1.3807103633880615,
+            1.3733720779418945,
+            1.3659957647323608,
+            1.3587980270385742,
+            1.3519399166107178,
+            1.3454982042312622,
+            1.3394557237625122,
+            1.3337373733520508,
+            1.3282841444015503,
+            1.323117971420288,
+            1.318347454071045,
+        ],
+    }
+    return loss_vs_iter.get(cell_id, [])
+
+
+def expected_mea_mape(cell_id: str):
+    """Below test data MEA and MAPE is generated with seed_everything(1)
+    Returns MEA and MAPE for given cell_id"""
+
+    MEA = {"Cell1": 1.7881388210634555, "Cell2": 0.07000514313754991, "Cell3": 0.07576626164573952}
+    MAPE = {"Cell1": 2.384185094751274, "Cell2": 0.07000514313754991, "Cell3": 0.09590666031106268}
+    return MEA.get(cell_id, []), MAPE.get(cell_id, [])
+
+
 def augment_ue_data(ue_data_df: pd.DataFrame, site_configs_df: pd.DataFrame):
     """
     This method takes in UE data and site config and adds to the UE data
@@ -196,93 +278,10 @@ def split_training_and_test_data(ue_data_df: pd.DataFrame, test_size: float):
     return cell_id_training_data_map, cell_id_test_data_map
 
 
-def seed_everything(seed: int):
-    random.seed(seed)
-    os.environ["PYTHONHASHSEED"] = str(seed)
-    np.random.seed(seed)
-    torch.manual_seed(seed)
-    torch.cuda.manual_seed(seed)
-    torch.backends.cudnn.deterministic = True
-    torch.backends.cudnn.benchmark = True
-
-
 class TestBayesianDigitalTwin(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         seed_everything(1)
-
-        # Test data loss_vs_iter generated with seed_everything(1)
-        cls.loss_vs_iter = {
-            "Cell1": [
-                1.3360857963562012,
-                1.3250646591186523,
-                1.3147950172424316,
-                1.305358648300171,
-                1.296657919883728,
-                1.288337230682373,
-                1.2800240516662598,
-                1.2716268301010132,
-                1.2632856369018555,
-                1.2552096843719482,
-                1.2475690841674805,
-                1.2404227256774902,
-                1.2336828708648682,
-                1.2271721363067627,
-                1.2207590341567993,
-                1.2144416570663452,
-                1.208317518234253,
-                1.202500820159912,
-                1.197044849395752,
-                1.1918997764587402,
-            ],
-            "Cell2": [
-                1.4958915710449219,
-                1.4850997924804688,
-                1.476243019104004,
-                1.46876859664917,
-                1.461909294128418,
-                1.455451250076294,
-                1.4494956731796265,
-                1.4441474676132202,
-                1.4394338130950928,
-                1.4352850914001465,
-                1.4315435886383057,
-                1.4280192852020264,
-                1.4245692491531372,
-                1.4211323261260986,
-                1.4177082777023315,
-                1.4143176078796387,
-                1.41096830368042,
-                1.4076392650604248,
-                1.404287576675415,
-                1.400879144668579,
-            ],
-            "Cell3": [
-                1.4421265125274658,
-                1.4319207668304443,
-                1.4234235286712646,
-                1.4165308475494385,
-                1.4108211994171143,
-                1.4056122303009033,
-                1.4002456665039062,
-                1.3943250179290771,
-                1.3877692222595215,
-                1.3807103633880615,
-                1.3733720779418945,
-                1.3659957647323608,
-                1.3587980270385742,
-                1.3519399166107178,
-                1.3454982042312622,
-                1.3394557237625122,
-                1.3337373733520508,
-                1.3282841444015503,
-                1.323117971420288,
-                1.318347454071045,
-            ],
-        }
-        # Test data MEA and MAPE generated with seed_everything(1)
-        cls.MEA = {"Cell1": 1.7881388210634555, "Cell2": 0.07000514313754991, "Cell3": 0.07576626164573952}
-        cls.MAPE = {"Cell1": 2.384185094751274, "Cell2": 0.07000514313754991, "Cell3": 0.09590666031106268}
 
     def test_rf_bayesian_digital_twin(self):
         # get sample data
@@ -324,11 +323,9 @@ class TestBayesianDigitalTwin(unittest.TestCase):
                 f" for cell {cell_id}, with min learning loss {min(loss_vs_iter):0.5f}, "
                 f"avg learning loss {np.mean(loss_vs_iter):0.5f} and final learning loss {loss_vs_iter[-1]:0.5f}"
             )
-            expected_loss_vs_iter = self.loss_vs_iter[cell_id]
-            self.assertEqual(loss_vs_iter.tolist(), expected_loss_vs_iter)
+            self.assertEqual(loss_vs_iter.tolist(), expected_loss_vs_iter(cell_id=cell_id))
 
         # predict/test
-
         for cell_id, testing_data in cell_id_test_data_map.items():
             (pred_means, _) = bayesian_digital_twin_map[cell_id].predict_distributed_gpmodel(
                 prediction_dfs=[testing_data]
@@ -340,8 +337,6 @@ class TestBayesianDigitalTwin(unittest.TestCase):
                 f"cell_id = {cell_id}, MAE = {MAE:0.5f} dB, MAPE = {MAPE:0.5f} %,"
                 "# test points = {len(testing_data.avg_rsrp)}"
             )
-            EXPECTED_MAE = self.MEA[cell_id]
+            EXPECTED_MAE, EXPECTED_MAPE = expected_mea_mape(cell_id=cell_id)
             self.assertEqual(MAE, EXPECTED_MAE)
-
-            EXPECTED_MAPE = self.MAPE[cell_id]
             self.assertEqual(MAPE, EXPECTED_MAPE)
